+++ src/serialization.cpp
@@ -240,7 +240,7 @@ size_t JsonWriteUnsubscribeCommand(char* dest, size_t maxLen, int nonce, const c
     return writer.Size();
 }
 
-size_t JsonWriteJoinReply(char* dest, size_t maxLen, const char* userId, int reply, int nonce)
+size_t JsonWriteJoinReply(char* dest, size_t maxLen, const char* userId, int8_t reply, int nonce)
 {
     JsonWriter writer(dest, maxLen);
