+++ src/serialization.h
@@ -66,7 +66,7 @@ size_t JsonWriteSubscribeCommand(char* dest, size_t maxLen, int nonce, const cha
 
 size_t JsonWriteUnsubscribeCommand(char* dest, size_t maxLen, int nonce, const char* evtName);
 
-size_t JsonWriteJoinReply(char* dest, size_t maxLen, const char* userId, int reply, int nonce);
+size_t JsonWriteJoinReply(char* dest, size_t maxLen, const char* userId, int8_t reply, int nonce);
 
 size_t JsonWriteAcceptInvite(char* dest,
                              size_t maxLen,
